logic fix, also add PlayAgainButton component
