+++ packages/ui/src/components/PlayAgainButton/index.ts
@@ -1,0 +1,1 @@
+export * from './PlayAgainButton';
