+++ packages/ui/src/components/index.ts
@@ -12,3 +12,4 @@ export * from './WalletBenefits';
 export * from './ProgressiveImg';
 export * from './Steps';
 export * from './CopyButton';
+export * from './PlayAgainButton';
